chore(objectify-words): clean files for PR
