+++ src/components/editor/text/Word.svelte
@@ -1,14 +1,7 @@
 <script lang="ts">
 	import type { WordI } from '$src/types';
 
-	export let word: WordI = {
-		string: '',
-		index: 0,
-		isWordiable: false,
-		color: 'black',
-		wordiablePos: null
-	};
-
+	export let word: WordI;
 	let color: string = word.color;
 
 	let text = ` ${word.string}`;
