+++ src/lib/editor.ts
@@ -1,63 +1,21 @@
-import { wordiables } from '$stores/text';
-import type { RegexI } from '$src/types';
+import { regex } from '$lib/regex';
 
-let matchedWords: string[] = [];
-
-const replaceNewlines = (str: string): string => {
-	return str.replace(/\n/g, ' <br> ');
-};
-
-const findWordiables = (text: string[]): string => {
-	return text.find((word) => {
-		const isNotInMatchedWords = !matchedWords.includes(word);
-		const hasOnlyTwoForwardSlashes = word.match(/\\/g).length === 2;
-		const isWordiable = isNotInMatchedWords && hasOnlyTwoForwardSlashes;
-		if (isWordiable) setWordiables(word);
-	});
-};
-
-const setWordiables = (word: string): never => {
-	wordiables.set([...matchedWords, word]);
-};
-
-const syncWordiables = (str: string[]): never => {
-	matchedWords.forEach((word) => {
-		if (!str.includes(word)) wordiables.set(matchedWords.filter((w) => w !== word));
-	});
-};
-
-const sortMatchedWords = (str: string[]): never => {
-	matchedWords.sort((a, b) => {
-		const aIndex = str.indexOf(a);
-		const bIndex = str.indexOf(b);
-		return aIndex - bIndex;
-	});
-};
-
-const syncMatches = (text: string[]): never => {
-	if (text.length < 1) return;
-	findWordiables(text);
-	syncWordiables(text);
-	sortMatchedWords(text);
-};
-
-export const regex: RegexI = {
-	wordiables: /\\(.*?)\\/g
+export const replaceNewlines = (str: string): string => {
+	return str.replace(regex.newline, ' <br> ');
 };
 
 export const space = '&nbsp;';
 
 export const splitText = (text: string): string[] => text.split(' ');
 
-export const checkForWordiables = (text: string): string => {
-	wordiables.subscribe((words) => (matchedWords = words));
-	const matches: string[] = text.match(regex.wordiables);
-	if (matches) syncMatches(matches);
-	return text;
+export const getStringFromText = (text: WordI[]): string => {
+	return text.reduce((acc: string, t: WordI): string => {
+		acc += ` ${t.string}`;
+		return acc;
+	}, '');
 };
 
-export const objectifyWord = (text) => {
-	const words = splitText(replaceNewlines(text));
+export const objectifyWords = (words: string[]): WordI => {
 	return words.reduce((acc, word, index) => {
 		const wordExpanded: WordI = {
 			string: replaceNewlines(word),
@@ -66,26 +24,7 @@ export const objectifyWord = (text) => {
 			color: 'black',
 			wordiablePos: null
 		};
-
-		acc = [...acc, wordExpanded];
-		console.log(acc);
+		acc.push(wordExpanded);
 		return acc;
 	}, []);
-};
-
-const rainbow: string[] = ['red', 'orange', 'yellow', 'green', 'blue', 'indigo', 'violet', 'black'];
-
-const isWordiableWithSlashes = (word: string): RegExpMatchArray | null =>
-	word.match(regex.wordiables);
-
-const isWordiable = (word: string): boolean => matchedWords.includes(word);
-
-export const powerWordiables = (text: object[]): never => {
-	text.forEach((t) => {
-		if (isWordiable(t.string) || isWordiableWithSlashes(t.string)) {
-			t.iswordiable = true;
-			t.wordiablePos = matchedWords.indexOf(t.string);
-			t.color = rainbow[t.wordiablePos];
-		}
-	});
 };
