+++ src/lib/regex.ts
@@ -1,0 +1,7 @@
+import type { RegexI } from '$src/types';
+
+export const regex: RegexI = {
+	wordiables: /\\(.*?)\\/g,
+	backslash: /\\/g,
+	newline: /\n/g
+};
