+++ src/stores/text.ts
@@ -1,12 +1,13 @@
 import { derived, writable } from 'svelte/store';
-
-import { objectifyWord, checkForWordiables } from '$src/lib/editor';
+import { objectifyWords, splitText, replaceNewlines } from '$src/lib/editor';
+import { checkForWordiables } from '$src/lib/wordiables';
 
 export const text = writable('');
 
 export const parsedText = derived(text, ($text) => {
 	checkForWordiables($text);
-	return objectifyWord($text);
+	const words = splitText(replaceNewlines($text));
+	return objectifyWords(words);
 });
 
 export const wordiables: never = writable([]);
